pyproject.toml
[tool.poetry]
name = "macos-stream-to-web"
version = "0.1.0"
description = "MacOS screen capturing"
authors = ["ansh <anshfreelance1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
websockets = "^12.0"
pyobjc-core = "^10.2"
pyobjc-framework-quartz = "^10.2"
pyobjc-framework-cocoa = "^10.2"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"
uvicorn = "^0.29.0"
jinja2 = "^3.1.3"
pillow = "^10.3.0"
ruff = "^0.4.3"
mypy = "^1.10.0"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
